Task 3

+++ app/src/main/java/com/aleksandersh/weather/network/dto/currentWeather/WindDto.java
@@ -13,13 +13,13 @@ public class WindDto {
     private Double mSpeed;
     @SerializedName("deg")
     @Expose
-    private Integer mDirection;
+    private Double mDirection;
 
     public Double getSpeed() {
         return mSpeed;
     }
 
-    public Integer getDirection() {
+    public Double getDirection() {
         return mDirection;
     }
 }
